refactor(video): extract video annotation model into vailabel-video crate
Phase 1 step 8c. Moves the video data model (Point/SceneCut/FrameThumb/Video/
Track/TrackKeyframe/MaterializedShape/VideoJob/FfmpegInfo/ProbeResult) into
crates/video domain and request payloads into contracts. now_iso in
VideoJob::new resolves vailabel_shared::now_iso; the module-doc super:: links
were delinked (don't resolve cross-crate). Video/Track impl core::Identifiable.
video/model.rs is a glob re-export shim; ffmpeg/interpolation/service stay in
the binary.
